file_client/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
grpcio==1.59.3
grpcio-tools==1.59.3
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
protobuf==4.25.1
pytest==7.4.3
requests==2.31.0
tomli==2.0.1
urllib3==2.1.0
